@@ -1,6 +1,6 @@
 // Tideland Go Audit - Capture - Unit Tests
 //
-// Copyright (C) 2017-2020 Frank Mueller / Tideland / Oldenburg / Germany
+// Copyright (C) 2017-2023 Frank Mueller / Tideland / Oldenburg / Germany
 //
 // All rights reserved. Use of this source code is governed
 // by the new BSD license.
