@@ -1,6 +1,6 @@
 // Tideland Go Audit - Web
 //
-// Copyright (C) 2012-2021 Frank Mueller / Tideland / Oldenburg / Germany
+// Copyright (C) 2012-2023 Frank Mueller / Tideland / Oldenburg / Germany
 //
 // All rights reserved. Use of this source code is governed
 // by the new BSD license.
@@ -9,17 +9,17 @@
 // with the handler to test. Then standard http requests can be sent
 // and the returned http responses can be analyzed.
 //
-//    h := NewMyHandler()
-//    s := web.NewSimulator(h)
+//	h := NewMyHandler()
+//	s := web.NewSimulator(h)
 //
-//    req := s.CreateRequest(http.MethodGet, "http://localhost:8080/", nil)
+//	req := s.CreateRequest(http.MethodGet, "http://localhost:8080/", nil)
 //
-//    resp, err := s.Do(req)
-//    assert.NoError(err)
-//    assert.Equal(resp.StatusCode, http.StatusOK)
-//    body, err := web.BodyToString(resp)
-//    assert.NoError(err)
-//    assert.Equal(body, test.expected)
+//	resp, err := s.Do(req)
+//	assert.NoError(err)
+//	assert.Equal(resp.StatusCode, http.StatusOK)
+//	body, err := web.BodyToString(resp)
+//	assert.NoError(err)
+//	assert.Equal(body, test.expected)
 //
 // Some smaller functions help working with the requests and responses.
 package web // import "tideland.dev/go/audit/web"
